+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/utils/BuiltinMimeTypes.kt
@@ -8,4 +8,11 @@ object BuiltinMimeTypes {
     object Video {
         val MP4 = buildMimeType("video/mp4")
     }
+    object Text {
+        val Html = buildMimeType("text/html")
+    }
+    object Application {
+        val Zip = buildMimeType("application/zip")
+        val Pdf = buildMimeType("application/pdf")
+    }
 }