+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/utils/ByteReadChannelToInput.kt
@@ -1,0 +1,6 @@
+package dev.inmo.tgbotapi.utils
+
+import io.ktor.utils.io.ByteReadChannel
+import io.ktor.utils.io.core.Input
+
+expect suspend fun ByteReadChannel.asInput(): Input