+++ tgbotapi.extensions.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/DownloadFile.kt
@@ -1,39 +1,33 @@
 package dev.inmo.tgbotapi.extensions.api
 
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.extensions.api.get.getFileAdditionalInfo
-import dev.inmo.tgbotapi.requests.DownloadFile
+import dev.inmo.tgbotapi.extensions.api.files.downloadFile
 import dev.inmo.tgbotapi.requests.abstracts.FileId
 import dev.inmo.tgbotapi.types.files.PathedFile
 import dev.inmo.tgbotapi.types.files.abstracts.TelegramMediaFile
 import dev.inmo.tgbotapi.types.message.content.abstracts.MediaContent
 
+@Deprecated("Replaced", ReplaceWith("downloadFile", "dev.inmo.tgbotapi.extensions.api.files.downloadFile"))
 suspend fun TelegramBot.downloadFile(
     filePath: String
-): ByteArray = execute(
-    DownloadFile(filePath)
-)
+): ByteArray = downloadFile(filePath)
 
+@Deprecated("Replaced", ReplaceWith("downloadFile", "dev.inmo.tgbotapi.extensions.api.files.downloadFile"))
 suspend fun TelegramBot.downloadFile(
     pathedFile: PathedFile
-): ByteArray = execute(
-    DownloadFile(pathedFile.filePath)
-)
+): ByteArray = downloadFile(pathedFile)
 
+@Deprecated("Replaced", ReplaceWith("downloadFile", "dev.inmo.tgbotapi.extensions.api.files.downloadFile"))
 suspend fun TelegramBot.downloadFile(
     fileId: FileId
-): ByteArray = downloadFile(
-    getFileAdditionalInfo(fileId)
-)
+): ByteArray = downloadFile(fileId)
 
+@Deprecated("Replaced", ReplaceWith("downloadFile", "dev.inmo.tgbotapi.extensions.api.files.downloadFile"))
 suspend fun TelegramBot.downloadFile(
     file: TelegramMediaFile
-): ByteArray = downloadFile(
-    getFileAdditionalInfo(file)
-)
+): ByteArray = downloadFile(file)
 
+@Deprecated("Replaced", ReplaceWith("downloadFile", "dev.inmo.tgbotapi.extensions.api.files.downloadFile"))
 suspend fun TelegramBot.downloadFile(
     file: MediaContent
-): ByteArray = downloadFile(
-    getFileAdditionalInfo(file.media)
-)
+): ByteArray = downloadFile(file)