+++ tgbotapi.extensions.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/get/GetFile.kt
@@ -4,6 +4,7 @@ import dev.inmo.tgbotapi.bot.TelegramBot
 import dev.inmo.tgbotapi.requests.abstracts.FileId
 import dev.inmo.tgbotapi.requests.get.GetFile
 import dev.inmo.tgbotapi.types.files.abstracts.TelegramMediaFile
+import dev.inmo.tgbotapi.types.message.content.abstracts.MediaContent
 
 suspend fun TelegramBot.getFileAdditionalInfo(
     fileId: FileId
@@ -14,3 +15,7 @@ suspend fun TelegramBot.getFileAdditionalInfo(
 suspend fun TelegramBot.getFileAdditionalInfo(
     file: TelegramMediaFile
 ) = getFileAdditionalInfo(file.fileId)
+
+suspend fun TelegramBot.getFileAdditionalInfo(
+    content: MediaContent
+) = getFileAdditionalInfo(content.media)